\name{bikesR-package}
\alias{bikesR-package}
\alias{bikesR}
\docType{package}
\title{
\packageTitle{bikesR}
}
\description{
\packageDescription{bikesR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bikesR}
\packageIndices{bikesR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bikesR}

Maintainer: \packageMaintainer{bikesR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
